# server/internal/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bastion API
  description: |
    API documentation for Bastion bot and webhook integrations.

    ## Authentication

    **Bot tokens** use the `Bot` scheme:
    ```
    Authorization: Bot bot_xxxxxxxxxxxx
    ```

    Bot tokens are issued when creating a bot in Server Settings > Integrations.
    They are shown once — copy and store securely. You can regenerate a token
    at any time (which invalidates the previous one).

    **Webhook tokens** are embedded in the URL and require no header:
    ```
    POST /api/v1/webhooks/{webhookID}/{token}
    ```

    **User JWTs** use the `Bearer` scheme (for management endpoints):
    ```
    Authorization: Bearer <jwt>
    ```

    ## WebSocket

    Bots can connect to the real-time WebSocket for live events:
    ```
    GET /api/v1/ws?token=bot_xxxxxxxxxxxx
    ```

    ## Rate Limits

    | Scope | Limit |
    |-------|-------|
    | Webhook execution | 30 requests/min per IP |
    | Protected API (global) | 120 requests/min per user |
    | Message send | 10 messages/10s per user |
    | File upload | 5 uploads/min per user |
    | Bulk import | 10 requests/min per user |

    When rate-limited, the server returns `429 Too Many Requests`.

  version: "1.0"
  contact:
    name: Bastion
    url: https://bastions.org
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /webhooks/{webhookID}/{token}:
    post:
      tags: [Webhooks]
      summary: Execute webhook
      description: |
        Send a message to the webhook's channel. This endpoint is public —
        no Authorization header is needed. The token in the URL authenticates
        the request.

        Rate limit: **30 requests/min** per IP.
      operationId: executeWebhook
      parameters:
        - $ref: '#/components/parameters/webhookID'
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: Webhook secret token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookExecuteRequest'
            example:
              content: "Build #142 passed"
              username: "CI Bot"
              avatarUrl: "https://example.com/ci-avatar.png"
              embeds:
                - title: "Build Report"
                  description: "All tests passed"
                  color: 65280
      responses:
        '200':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          description: Rate limited

  # ── Webhook CRUD ────────────────────────────────────────────────────
  /servers/{serverID}/webhooks:
    post:
      tags: [Webhooks]
      summary: Create webhook
      description: Creates a webhook for a channel. Requires `ManageServer` permission.
      operationId: createWebhook
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
            example:
              name: "GitHub Notifications"
              channelId: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '201':
          description: Webhook created (includes token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'

    get:
      tags: [Webhooks]
      summary: List webhooks
      description: Lists all webhooks for a server. Tokens are hidden. Requires `ManageServer` permission.
      operationId: listWebhooks
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
      responses:
        '200':
          description: Webhook list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Webhook'
  /servers/{serverID}/webhooks/{webhookID}:
    get:
      tags: [Webhooks]
      summary: Get webhook
      description: Get a single webhook including its token. Requires `ManageServer` permission.
      operationId: getWebhook
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
        - $ref: '#/components/parameters/webhookID'
      responses:
        '200':
          description: Webhook details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Webhooks]
      summary: Update webhook
      description: Update a webhook's name or channel. Requires `ManageServer` permission.
      operationId: updateWebhook
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
        - $ref: '#/components/parameters/webhookID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWebhookRequest'
      responses:
        '200':
          description: Webhook updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Webhooks]
      summary: Delete webhook
      description: Delete a webhook. Requires `ManageServer` permission.
      operationId: deleteWebhook
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
        - $ref: '#/components/parameters/webhookID'
      responses:
        '200':
          description: Webhook deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Bot CRUD ────────────────────────────────────────────────────────
  /servers/{serverID}/bots:
    post:
      tags: [Bots]
      summary: Create bot
      description: |
        Creates a bot for the server. The bot token is returned **once** in the
        response — store it securely. Requires `ManageServer` permission.
      operationId: createBot
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBotRequest'
            example:
              username: "ModBot"
              description: "Automated moderation assistant"
      responses:
        '201':
          description: Bot created (includes token — shown once)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bot'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'

    get:
      tags: [Bots]
      summary: List bots
      description: Lists all bots in the server. Requires `ManageServer` permission.
      operationId: listBots
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
      responses:
        '200':
          description: Bot list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bot'
  /servers/{serverID}/bots/{botID}:
    get:
      tags: [Bots]
      summary: Get bot
      description: Get a single bot's details. Requires `ManageServer` permission.
      operationId: getBot
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
        - $ref: '#/components/parameters/botID'
      responses:
        '200':
          description: Bot details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bot'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Bots]
      summary: Update bot
      description: Update a bot's username or description. Requires `ManageServer` permission.
      operationId: updateBot
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
        - $ref: '#/components/parameters/botID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBotRequest'
      responses:
        '200':
          description: Bot updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bot'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Bots]
      summary: Delete bot
      description: Delete a bot. Requires `ManageServer` permission.
      operationId: deleteBot
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
        - $ref: '#/components/parameters/botID'
      responses:
        '200':
          description: Bot deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /servers/{serverID}/bots/{botID}/regenerate-token:
    post:
      tags: [Bots]
      summary: Regenerate bot token
      description: |
        Generates a new token for the bot. The old token is immediately
        invalidated. The new token is returned **once** — store it securely.
        Requires `ManageServer` permission.
      operationId: regenerateBotToken
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
        - $ref: '#/components/parameters/botID'
      responses:
        '200':
          description: New token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "bot_abc123def456..."
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Application Commands (Interactions) ──────────────────────────────
  /servers/{serverID}/bots/{botID}/commands:
    post:
      tags: [Interactions]
      summary: Register command
      description: |
        Register a slash command (or context menu command) for a bot.
        Requires bot token auth (bot must belong to the server).
      operationId: registerCommand
      security:
        - BotAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
        - $ref: '#/components/parameters/botID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommandRequest'
            example:
              name: "ping"
              description: "Responds with pong"
              type: 1
      responses:
        '201':
          description: Command created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationCommand'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Command with that name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags: [Interactions]
      summary: List bot commands
      description: List all commands registered by a specific bot.
      operationId: listBotCommands
      security:
        - BotAuth: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
        - $ref: '#/components/parameters/botID'
      responses:
        '200':
          description: Command list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicationCommand'
  /servers/{serverID}/bots/{botID}/commands/{commandID}:
    patch:
      tags: [Interactions]
      summary: Update command
      description: Update a command's description or options.
      operationId: updateCommand
      security:
        - BotAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
        - $ref: '#/components/parameters/botID'
        - name: commandID
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Application command ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommandRequest'
      responses:
        '200':
          description: Command updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationCommand'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Interactions]
      summary: Delete command
      description: Delete a registered command.
      operationId: deleteCommand
      security:
        - BotAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
        - $ref: '#/components/parameters/botID'
        - name: commandID
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Application command ID
      responses:
        '200':
          description: Command deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /servers/{serverID}/commands:
    get:
      tags: [Interactions]
      summary: List server commands
      description: |
        List all slash commands and context menu commands available in a server
        (from all bots). Requires server membership.
      operationId: listServerCommands
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
      responses:
        '200':
          description: All commands in the server
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicationCommand'
  /servers/{serverID}/interactions:
    post:
      tags: [Interactions]
      summary: Execute interaction
      description: |
        Invoke a slash command or context menu command. The server creates a
        short-lived interaction token and dispatches an `INTERACTION_CREATE`
        event to the bot via WebSocket. The bot then responds via the callback
        endpoint.
      operationId: executeInteraction
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteInteractionRequest'
            example:
              commandId: "550e8400-e29b-41d4-a716-446655440000"
              channelId: "550e8400-e29b-41d4-a716-446655440001"
      responses:
        '204':
          description: Interaction dispatched to bot
        '404':
          $ref: '#/components/responses/NotFound'
  /interactions/{token}/callback:
    post:
      tags: [Interactions]
      summary: Interaction callback
      description: |
        Bot responds to an interaction. The token is from the
        `INTERACTION_CREATE` event payload. Tokens expire after 15 minutes.

        If `ephemeral: true`, the response is sent only to the invoking user
        via WebSocket and is **not** stored in the database.

        This endpoint is public — authenticated by the token in the URL.
      operationId: interactionCallback
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: Interaction token from the INTERACTION_CREATE event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionCallbackRequest'
            example:
              content: "Pong!"
              ephemeral: false
      responses:
        '204':
          description: Response delivered
        '404':
          description: Token not found or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Messages ─────────────────────────────────────────────────────────
  /channels/{channelID}/messages:
    get:
      tags: [Messages]
      summary: List messages
      description: |
        Fetch messages in a channel. Supports cursor-based pagination via the
        `before` query parameter. Returns up to 50 messages per request.
      operationId: listMessages
      security:
        - BotToken: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/channelID'
        - name: before
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Fetch messages before this message ID (cursor pagination)
      responses:
        '200':
          description: Message list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'

    post:
      tags: [Messages]
      summary: Send message
      description: |
        Send a message to a channel. Must have either `content` or `embeds` (or both).

        Rate limit: **10 messages per 10 seconds**.

        **Bot-only features:**
        - `createdAt` — backdate the message (useful for historical import)
        - `embeds` — rich embed objects
      operationId: sendMessage
      security:
        - BotToken: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/channelID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
            example:
              content: "Hello from my bot!"
              embeds:
                - title: "Status Update"
                  description: "Server is online"
                  color: 65280
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          description: Rate limited
  /channels/{channelID}/messages/{messageID}:
    put:
      tags: [Messages]
      summary: Edit message
      description: |
        Edit a message. Only the message author can edit it.
        Must have either `content` or `embeds` (or both).
      operationId: editMessage
      security:
        - BotToken: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/channelID'
        - $ref: '#/components/parameters/messageID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditMessageRequest'
      responses:
        '200':
          description: Message updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Messages]
      summary: Delete message
      description: |
        Delete a message. Authors can delete their own messages.
        Users with `ManageMessages` permission can delete any message.
      operationId: deleteMessage
      security:
        - BotToken: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/channelID'
        - $ref: '#/components/parameters/messageID'
      responses:
        '200':
          description: Message deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /servers/{serverID}/members:
    get:
      tags: [Members]
      summary: List members
      description: List all members in a server. Bot must be a member of the server.
      operationId: listMembers
      security:
        - BotToken: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
      responses:
        '200':
          description: Member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Member'
...
components:
  securitySchemes:
    BotToken:
      type: apiKey
      in: header
      name: Authorization
      description: 'Bot token: `Bot bot_xxxxxxxxxxxx`'
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from login
  parameters:
...
    webhookID:
      name: webhookID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Webhook ID
...
  schemas:
...
    WebhookExecuteRequest:
      type: object
      properties:
        content:
          type: string
          maxLength: 4000
          description: Message content (required if no embeds)
        username:
          type: string
          description: Override the webhook's display name for this message (persisted as authorOverride)
        avatarUrl:
          type: string
          format: uri
          description: Override the webhook's avatar for this message (persisted as authorOverride)
        embeds:
          type: array
          items:
            $ref: '#/components/schemas/Embed'
          maxItems: 10
          description: Rich embed objects (required if no content)
...